+++ packages/sv/src/addons/common.ts
@@ -1,6 +1,6 @@
+import process from 'node:process';
 import { log } from '@clack/prompts';
 import { color, dedent, type SvelteAst, type TransformFn, transforms } from '@sveltejs/sv-utils';
-import process from 'node:process';
 
 // This is in common because the eslint addon installs this version,
 // and the prettier addon uses this to check if the installed major version of